requirements.txt
langchain-pinecone
pinecone-notebooks
langchain_community
langchain-huggingface
langchain-text-splitters
langchain-groq
flask
pypdf
python-dotenv
-e .